+++ test/unit/resolveAllOf.test.ts
@@ -316,6 +316,7 @@ describe("resolveAllOf", () => {
                                 }
                             },
                             then: {
+                                required: ["additionalSchema"],
                                 properties: {
                                     additionalSchema: { type: "string", default: "additional" }
                                 }
@@ -329,6 +330,7 @@ describe("resolveAllOf", () => {
                                 }
                             },
                             then: {
+                                required: ["anotherSchema"],
                                 properties: {
                                     anotherSchema: { type: "string", default: "another" }
                                 }
@@ -339,7 +341,7 @@ describe("resolveAllOf", () => {
             );
             expect(schema).to.deep.equal({
                 type: "object",
-                required: ["trigger"],
+                required: ["trigger", "additionalSchema", "anotherSchema"],
                 properties: {
                     trigger: { type: "boolean" },
                     additionalSchema: { type: "string", default: "additional" },